Add styled links, move padding, move colors
This is to cut down a lot of the inline CSS, and to add some user friendly functionality.

I moved background color and text color to the style.css file. Anything with consistent padding was also moved like lists and the such. Added two new user friendly links. One home "button" and a scroll to top "button". They are just stylized links.